Improve Helm configuration
The current prelude-helm only uses a single command of Helm, which
is a waste given how Helm contains many other commands. When
prelude-helm is activated, users should be able to utilize all of these
useful commands. If a user wants to use Helm, he will want to use all of
its features anyway.

prelude-helm is configured that it is able to be used with this [guide](http://tuhdo.github.io/helm-intro.html).