Replace usages of eval-after-load with with-eval-after-load
`with-eval-after-load` was introduced in Emacs 24.4 and its
usage results in a cleaner code.

`eval-after-load` is considered ill-behaved because it is a function,
not a macro, and thus requires the code inside it to be quoted, which
means that it cannot be byte-compiled. It also accepts only one form,
so if you have more than one, you need to use `progn`.

More details - https://stackoverflow.com/questions/21880139/what-is-with-eval-after-load-in-emacs-lisp